Todo funciona bien hasta aqui

+++ src/main/java/myApp/blog/repositorio/RecetaRepositorio.java
@@ -3,5 +3,5 @@ package myApp.blog.repositorio;
 import myApp.blog.modelo.Receta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecetaRepositorio extends JpaRepository<Receta, Integer> {
+public interface RecetaRepositorio extends JpaRepository<Receta, Long> {
 }
